Lower the priority of tracker related mixins with overwrites.
